+++ application/controllers/Assembly.php
@@ -29,7 +29,7 @@ class Assembly extends Application
                 case "Assemble":
                     if (count($this->input->post('1')) != 1
                         || count($this->input->post('2')) != 1
-                        || count($this->input->post('3')) !=1) 
+                        || count($this->input->post('3')) !=1)
                     {
                         echo "Incorrect number of parts";
                         break;
@@ -37,38 +37,38 @@ class Assembly extends Application
                     $parts[] = $this->parts->get($this->input->post('1')[0]);
                     $parts[] = $this->parts->get($this->input->post('2')[0]);
                     $parts[] = $this->parts->get($this->input->post('3')[0]);
-                    
+
                     //create empty objects
                     $robot = new StdClass;
                     $transaction = new StdClass;
-                    
+
                     //add parts to the robot
                     $robot->id = "";
                     $robot->head = $parts[0]->certificate;
                     $robot->torso = $parts[1]->certificate;
                     $robot->legs = $parts[2]->certificate;
-                    
+
                     $this->robots->add($robot);
-                    
+
                     //add transaction to the history
                     $transaction->id = "";
                     $transaction->type = "robot_build";
                     $transaction->part_id = "";
                     $transaction->robot_id = $this->robots->highest();
                     $transaction->amount = 0;
-                    
+
                     $this->history->add($transaction);
-                    
+
                     //make parts no longer available
                     foreach($parts as $part) {
                         $part->available = 0;
                         $this->parts->update($part);
                     }
-                    
+
                     echo "Assembled";
-                    
+
                     break;
-                    
+
                 //return the parts to the server
                 case "Return":
                     if (count($this->input->post('1')) > 0)
@@ -80,10 +80,10 @@ class Assembly extends Application
                     if (count($this->input->post('3')) > 0)
                         foreach($this->input->post('3') as $part)
                             $parts[] = $part;
-                    
+
                     if (count($parts) > 3 || count($parts) == 0)
                         break;
-                        
+
                     $result = $this->umbrella->recycle($parts);
 
                     if ($result['OK'] == 1) {
@@ -91,28 +91,28 @@ class Assembly extends Application
                             $transaction = new StdClass;//create empty trasaction
                             //add transaction to the history
                             $transaction->id = "";
-                            $transaction->type = "part_sale";
+                            $transaction->type = "part_recycle";
                             $transaction->part_id = $part;
                             $transaction->robot_id = "";
                             $transaction->amount = $result['response'] / count($parts);
-                            
+
                             $this->history->add($transaction);
-                            
+
                             $this->parts->delete($part);
                         }
                     } else {
                         break;
                     }
-                    
+
                     echo "Returned, " . "$" . $result['response'];
-                    
+
                     break;
                 default:
                     break;
             }
             redirect($_SERVER['HTTP_REFERER']); //back where we came from
         }
-        
+
         /**
          * Ships the robot selected in the 'robot' set of radio buttons on the
          * view.  Then creates a transaction, and deletes the robot from the DB
@@ -122,37 +122,37 @@ class Assembly extends Application
                 return;
             if ($this->input->post('robot') == null)
                 return;
-            
+
             //setup parts into an array
             $robot = $this->robots->get($this->input->post('robot'));
             $parts[] = $robot->head;
             $parts[] = $robot->torso;
             $parts[] = $robot->legs;
-            
+
             //try to sell parts
             $result = $this->umbrella->buymybot($parts);
             if ($result['OK'] != 1) {
                 return;
             }
-            
+
             $transaction = new StdClass;//create empty trasaction
             //add transaction to the history
             $transaction->id = "";
-            $transaction->type = "part_sale";
+            $transaction->type = "robot_sale";
             $transaction->part_id = "";
             $transaction->robot_id = $robot->id;
             $transaction->amount = $result['response'] / count($parts);
-            
+
             $this->history->add($transaction);
-            
+
             //if successful, delete robot
             $this->robots->delete($robot->id);
-            
+
             echo "Shipped, " . "$" . $result['response'];
-            
+
             redirect($_SERVER['HTTP_REFERER']); //back where we came from
         }
-        
+
 	/**
 	 * Index Page for this controller.
 	 *
@@ -168,7 +168,7 @@ class Assembly extends Application
 	public function index()
 	{
 		$this->data['pagebody'] = 'assembly_view';
-                
+
                 //get all robots
                 $sourceRobots = $this->robots->all();
                 $robots = array ();
@@ -177,7 +177,7 @@ class Assembly extends Application
                     $head = $this->parts->get($robot->head);
                     $torso = $this->parts->get($robot->torso);
                     $legs = $this->parts->get($robot->legs);
-                    
+
                     $robots[] = array ('id' => $robot->id,
                         'head' => $robot->head,
                         'headImage' => $head->model . $head->piece . '.jpeg',